feat: add test generator panel
